Feat : separate type & logic
