+++ src/components/manage-task-item/ManageTaskItem.tsx
@@ -1,139 +1,31 @@
 'use client';
 
-import { useState } from 'react';
 import clsx from 'clsx';
-import { format } from 'date-fns';
 import FormField from '../common/formField';
 import { OptionSelector } from '../common/dropdown/OptionSelector';
 import CalendarSelect from '../calendar/CalendarSelect';
 import WeeklySelect from './WeeklySelect';
 import TimePicker from './TimePicker';
-import generateTime from './time-table';
-import {
-  Frequency,
-  Task,
-} from '@/app/(content-layout)/[groupId]/tasklist/_tasklist/types/task-list-page-type';
-
-interface TaskItem extends Pick<Task, 'id' | 'name' | 'frequency' | 'weekDays' | 'description'> {
-  date: Date | string;
-}
-
-export interface Time {
-  period: '오전' | '오후';
-  time: string;
-}
+import { TaskItem } from './type';
+import useManageTaskItem from './useManageTaskItem';
 
 const FREQUENCY_LIST = ['한 번', '매일', '주 반복', '월 반복'];
 
-const INITIAL_TASK_ITEM: TaskItem = {
-  id: 0,
-  name: '',
-  description: '',
-  date: new Date(),
-  frequency: 'ONCE',
-  weekDays: [],
-};
-
-const FREQUENCY_MAP: Record<Frequency, string> = {
-  ONCE: '한 번',
-  DAILY: '매일',
-  WEEKLY: '주 반복',
-  MONTHLY: '월 반복',
-};
-
-const REVERSE_FREQUENCY_MAP: Record<string, Frequency> = {
-  '한 번': 'ONCE',
-  매일: 'DAILY',
-  '주 반복': 'WEEKLY',
-  '월 반복': 'MONTHLY',
-};
-
 export default function ManageTaskItem({ task }: { task?: TaskItem }) {
-  const { am, pm } = generateTime();
-
-  const [isCalendarOpen, setIsCalendarOpen] = useState(false);
-  const [taskItem, setTaskItem] = useState<TaskItem>(() => ({
-    ...INITIAL_TASK_ITEM,
-    ...task,
-  }));
-  const [isTimeOpen, setIsTimeOpen] = useState(false);
-  const [selectedFrequency, setSelectedFrequency] = useState(() => {
-    return task?.frequency ? FREQUENCY_MAP[task.frequency] : '';
-  });
-  const [selectedTime, setSelectedTime] = useState<Time>({
-    period: '오전',
-    time: am[0],
-  });
-
-  const select = [
-    {
-      id: 'date',
-      value: format(taskItem.date, 'yyyy년 MM월 dd일'),
-      onClick: () => {
-        setIsCalendarOpen((prev) => !prev);
-        setIsTimeOpen(false);
-      },
-      isOpen: isCalendarOpen,
-      flex: 'flex-[1.65]',
-    },
-    {
-      id: 'time',
-      value: `${selectedTime.period} ${selectedTime.time}`,
-      onClick: () => {
-        setIsTimeOpen((prev) => !prev);
-        setIsCalendarOpen(false);
-      },
-      isOpen: isTimeOpen,
-      flex: 'flex-[1]',
-    },
-  ];
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement | HTMLTextAreaElement>) => {
-    const { name, value } = e.target;
-    setTaskItem((prev) => ({
-      ...prev,
-      [name]: value,
-    }));
-  };
-
-  const handleCalendarDateChange = (selectedDate: Date) => {
-    setTaskItem((prev) => ({
-      ...prev,
-      date: selectedDate,
-    }));
-    setIsCalendarOpen(false);
-  };
-
-  const handleFrequencyChange = (e: React.MouseEvent<HTMLDivElement>) => {
-    const frequency = e.currentTarget.textContent ?? '';
-    setSelectedFrequency(frequency);
-    setTaskItem((prev) => ({
-      ...prev,
-      frequency: REVERSE_FREQUENCY_MAP[frequency] || 'ONCE',
-    }));
-  };
-
-  const toggleDay = (idx: number) => {
-    setTaskItem((prev) => ({
-      ...prev,
-      weekly: prev.weekDays.includes(idx)
-        ? prev.weekDays.filter((i) => i !== idx)
-        : [...prev.weekDays, idx],
-    }));
-  };
-
-  const updateTime = (key: 'period' | 'time', value: string) => {
-    setSelectedTime((prev) => {
-      if (key === 'period') {
-        const newTime = value === '오전' ? am[0] : pm[0];
-        return { period: value as '오전' | '오후', time: newTime };
-      }
-
-      setIsTimeOpen(false);
-
-      return { ...prev, [key]: value };
-    });
-  };
+  const {
+    taskItem,
+    selectedTime,
+    isWeekly,
+    isCalendarOpen,
+    isTimeOpen,
+    select,
+    FREQUENCY_MAP,
+    handleInputChange,
+    handleCalendarDateChange,
+    handleFrequencyChange,
+    toggleDay,
+    updateTime,
+  } = useManageTaskItem({ task });
 
   const headingText = task ? '수정하기' : '만들기';
 
@@ -194,7 +86,7 @@ export default function ManageTaskItem({ task }: { task?: TaskItem }) {
           />
         </div>
 
-        {selectedFrequency === '주 반복' && (
+        {isWeekly && (
           <WeeklySelect
             selectedDays={taskItem.weekDays}
             toggleDay={(idx: number) => toggleDay(idx)}
